Refactoring MIDI Learn
git-svn-id: https://svn.code.sf.net/p/fsthost/code/trunk@292 75a15a69-0cfc-47d9-a1bd-ce2941b697b2

+++ jackvst.h
@@ -73,9 +73,9 @@ typedef struct _JackVST {
 
     int32_t         midi_map[128];
     bool            midi_learn;
-    short           midi_learn_CC;
-    int             midi_learn_PARAM;
-    short           midi_pc;
+    int8_t          midi_learn_CC;
+    int32_t         midi_learn_PARAM;
+    int8_t          midi_pc;
     MIDIFILTER*     filters;
     MIDIFILTER*     transposition;
     OCH_FILTERS     channel;
